test/test_agent/requirements.txt
# Requirements for Sleep Consultation Agent Testing Framework

# Core dependencies
langchain-openai>=0.1.0
langchain-core>=0.2.0
langgraph>=0.1.0
openai>=1.0.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlite3  # Built-in with Python

# Environment management
python-dotenv>=1.0.0

# Optional: For enhanced analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0